requirements.txt
appdirs==1.4.4
astroid==2.6.2
black==21.6b0
click==8.0.1
greenlet==1.1.0
isort==5.9.1
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pylint==2.9.3
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
regex==2021.7.6
SQLAlchemy==1.4.20
toml==0.10.2
Werkzeug==2.0.1
wrapt==1.12.1
